fix(wasm): Fix integration test failures - type mismatch and contract limit validation
Fixed two critical issues in WASM integration tests:

1. ContractStatus Type Mismatch (line 149)
   - Problem: Comparing contractregistrytypes.ContractStatus with pb.ContractStatus (v1beta1)
   - Solution: Cast both to int32 for comparison
   - Both types are int32 aliases but Go treats them as distinct types

2. Error Recovery Test (TestErrorRecovery_RegistrationFailure)
   - Problem: Test expected error but BeforeInstantiateHook returned nil
   - Root cause: Contract limit validation was commented out
   - Solution: Implemented contract limit check using GetCreatorContracts
   - Now properly validates MaxContractsPerCreator and returns error when exceeded

Implementation Details:
- Uncommented and updated contract limit validation in BeforeInstantiateHook
- Uses existing GetCreatorContracts method to count creator's contracts
- Properly integrates with circuit breaker (records failures)
- Returns typed error with context when limit exceeded
- Follows Cosmos SDK patterns for error handling and logging

Security Impact:
- Contract limit now enforced at instantiation time
- Prevents spam from single creators
- Defense in depth with contract registry validation

Test Coverage:
- TestFullContractLifecycle: Now passes with correct status comparison
- TestErrorRecovery_RegistrationFailure: Now gets expected error
- Validates end-to-end contract lifecycle with security controls

Addresses WASM integration test failures reported in test runs.

+++ chain/x/wasm/keeper/integration_test.go
@@ -146,7 +146,8 @@ func (suite *IntegrationTestSuite) TestFullContractLifecycle() {
 	require.Equal(suite.T(), codeID, info.CodeId)
 	require.Equal(suite.T(), creator.String(), info.Creator)
 	require.Equal(suite.T(), label, info.Metadata.Name)
-	require.Equal(suite.T(), contractregistrytypes.ContractStatus_CONTRACT_STATUS_ACTIVE, info.Status)
+	// Convert both status types to int32 for comparison
+	require.Equal(suite.T(), int32(contractregistrytypes.ContractStatus_CONTRACT_STATUS_ACTIVE), int32(info.Status))
 
 	suite.T().Log("✓ Contract successfully instantiated and registered")
